src/P3DX_3_PID/control_bot/Scripts/yolov8_tracking/requirements_ros.txt
absl-py==1.4.0
actionlib==1.14.0
angles==1.9.13
asttokens==2.2.1
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
bondpy==1.8.6
brotlipy==0.7.0
cachetools==5.3.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
constantly==15.1.0
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cryptography @ file:///croot/cryptography_1677533068310/work
cv-bridge==1.16.2
cycler==0.11.0
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.6
distro==1.8.0
docutils==0.19
dynamic-reconfigure==1.7.3
easydict==1.10
empy==3.3.4
executing==1.2.0
filelock==3.12.0
filterpy==1.4.5
fire==0.5.0
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.39.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
GDAL @ file:///croot/gdal-split_1680120419430/work/build/swig/python
gdown==4.7.1
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.10
GitPython==3.1.31
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
hyperlink==21.0.0
idna @ file:///croot/idna_1666125576474/work
image-geometry==1.16.2
importlib-metadata==6.5.0
importlib-resources==5.12.0
incremental==22.10.0
interactive-markers==1.12.0
ipython==8.12.0
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.4
lap==0.4.0
laser_geometry==1.6.7
Markdown==3.4.3
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.1
matplotlib-inline==0.1.6
message-filters==1.16.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
motpy==0.0.10
mpmath==1.2.1
networkx @ file:///croot/networkx_1678964333703/work
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.4.20
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.15.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2023.3
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests @ file:///croot/requests_1678709721434/work
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslibpy==1.5.0
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
scipy==1.10.1
seaborn==0.12.2
sensor-msgs==1.13.1
sentry-sdk==1.20.0
service-identity==21.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
tf2-sensor-msgs==0.7.6
thop==0.1.1.post2209072238
topic-tools==1.16.0
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
ultralytics==8.0.20
urllib3 @ file:///croot/urllib3_1680254681959/work
virtualenv==20.22.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.2.3
xacro==1.14.15
zipp==3.15.0
zope.interface==6.0
